backend/requirements.txt
# FastAPI
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic-settings==2.1.0

# MongoDB
motor==3.3.2
pymongo==4.6.1

# Vector Database & Embeddings
chromadb==0.4.18
sentence-transformers==2.2.2

# LLM - Groq (FREE)
groq==0.4.2

# File Processing
pypdf==3.17.1
python-pptx==0.6.23
pytesseract==0.3.10
Pillow==10.1.0
youtube-transcript-api==0.6.1
aiofiles==23.2.1

# Text Processing
langchain==0.0.350
tiktoken==0.5.2

# Utilities
python-dotenv==1.0.0
httpx==0.25.2
